pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "django-logger-extra"
authors = [{ name = "City of Helsinki", email = "dev@hel.fi" }]
description = "A module that provides extras for logger."
readme = "README.md"
version = "0.0.0"
keywords=[
    "logging",
    "extra tools",
    "plugin extension",
]
classifiers=[
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
license = "MIT"
license-files = ["LICENSE"]
dependencies = ["django>=4.2",]

[project.optional-dependencies]
extras = [
    "django-auditlog"
]
test = [
    "pytest",
    "pytest-django",
    "coverage-enable-subprocess",
    "coverage[toml]",
    "pytest-randomly",
    "pytest-rerunfailures",
    "pytest-xdist[psutil]",
]
auditlog = ["django-auditlog"]
all = ["django-logger-extra[auditlog]", "django-logger-extra[test]"]

[project.urls]
Homepage = "https://github.com/City-of-Helsinki/django-logger-extra"
Issues = "https://github.com/City-of-Helsinki/django-logger-extra/issues"

[tool.hatch.build.targets.wheel]
packages = ["logger_extra"]

[tool.hatch.version]
path = "logger_extra/__init__.py"

[tool.hatch.envs.default]
dependencies = [
    "pre-commit",
]
features = ["all"]
python = "3.9"
type = "virtual"
path = ".venv"

[tool.hatch.envs.default.scripts]
test = "pytest {args}"
lint = [
    "pre-commit install",
    "pre-commit run -a",
]
manage = "tests/manage.py {args}"

[tool.hatch.envs.hatch-test]
default-args = ["tests"]
features = ["all"]
matrix-name-format = "dj{value}"
parallel = false
randomize = true

[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.9"]
django = ["4.2"]

[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.10", "3.11", "3.12"]
django = ["4.2", "5.0", "5.1", "5.2"]

[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.13"]
django = ["5.1", "5.2"]

[tool.hatch.envs.hatch-test.overrides]
matrix.django.dependencies = [
    { value = "django~={matrix:django}" },
]

[tool.hatch.envs.hatch-test.scripts]
run = [
    "pytest{env:HATCH_TEST_ARGS:} {args}",
]
run-cov = "coverage run -m pytest{env:HATCH_TEST_ARGS:} {args}"
cov-combine = "coverage combine"
cov-report = "coverage report"

[tool.pytest.ini_options]
django_find_project = false
testpaths = [
    "tests",
]
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = "."

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # pycodestyle
    "E",
    "W",
    # isort
    "I",
    # pep8-naming
    "N",
    # flake8-bugbear without opinionated rules
    "B0",
    # flake8-pie
    "PIE",
    # flake8-print
    "T20",
]
extend-per-file-ignores = { "*/migrations/*" = ["E501"], "*/tests/*" = ["E501"] }

[tool.coverage.run]
source = ["logger_extra"]
omit = ["logger_extra/migrations/*", "*/tests/*"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "tests.settings"
